Refactor available_min_user_levels to view helper.
